+++ src/transformers/__init__.py
@@ -170,7 +170,7 @@ if is_sklearn_available():
 # Modeling
 if is_torch_available():
     from .modeling_utils import PreTrainedModel, prune_layer, Conv1D, apply_chunking_to_forward
-    from .modeling_generation import top_k_top_p_filtering
+    from .modeling_generation_utils import top_k_top_p_filtering
     from .modeling_auto import (
         AutoModel,
         AutoModelForPreTraining,
@@ -407,7 +407,7 @@ if is_torch_available():
 
 # TensorFlow
 if is_tf_available():
-    from .modeling_tf_generation import (
+    from .modeling_tf_generation_utils import (
         shape_list,
         tf_top_k_top_p_filtering,
     )
